src/main/java/com/example/demo/controller/MessageController.java
package com.example.demo.controller;

import java.util.Map;

import com.example.demo.service.MessageService;
import lombok.AccessLevel;
import lombok.experimental.FieldDefaults;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;


@RestController
@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
public class MessageController {

    MessageService messageService;

    @Autowired
    public MessageController(MessageService messageService) {
        this.messageService = messageService;
    }

    @PostMapping("/send")
    public ResponseEntity<Map<Character, Integer>> method(@RequestBody String str) {
        return ResponseEntity.ok(messageService.method(str));
    }

}
